infrastructure: Made weekPro and DailyObj and WeekDots viewModels seperate
